Removed duplicate "SETMessage" in "Link Name" Option and hardcoded obstruction for the Owl Options
I removed the ID 1001 SetMessage duplicate and had to remove a hardcoded gravestone in one of "Remove Owl" options.

This object prevented the player to get access to the hole with the three Deku salesmen as a child, due to hardcoded relation with the Lake Hylia Owl, which stayed unmovable with the actor gone.

It's only for the Child Link's version of the map. Adult's one is fine.